Renames and other changes
