+++ kermit-bugsnag/src/androidMain/kotlin/co/touchlab/kermit/bugsnag/BugsnagLogger.kt
@@ -10,7 +10,7 @@
 
 package co.touchlab.kermit.bugsnag
 
-import co.touchlab.kermit.Logger
+import co.touchlab.kermit.LogWriter
 import co.touchlab.kermit.Severity
 import com.bugsnag.android.Bugsnag
 
@@ -18,7 +18,7 @@ actual class BugsnagLogger actual constructor(
     private val minSeverity: Severity,
     private val minCrashSeverity: Severity,
     private val printTag: Boolean
-) : Logger() {
+) : LogWriter() {
     init {
         assert(minSeverity <= minCrashSeverity) {
             "minSeverity ($minSeverity) cannot be greater than minCrashSeverity ($minCrashSeverity)"
