+++ kermit-bugsnag/src/commonMain/kotlin/co/touchlab/kermit/bugsnag/BugsnagLogger.kt
@@ -10,11 +10,11 @@
 
 package co.touchlab.kermit.bugsnag
 
-import co.touchlab.kermit.Logger
+import co.touchlab.kermit.LogWriter
 import co.touchlab.kermit.Severity
 
 expect class BugsnagLogger(
     minSeverity: Severity = Severity.Info,
     minCrashSeverity: Severity = Severity.Warn,
     printTag: Boolean = true
-) : Logger
+) : LogWriter
