+++ kermit/src/commonMain/kotlin/co/touchlab/kermit/LockMutableKermitConfig.kt
@@ -15,8 +15,8 @@ import co.touchlab.stately.concurrency.withLock
 
 internal class LockMutableKermitConfig : MutableKermitConfig {
     private var _minSeverity: Severity = Severity.Debug
-    private var _loggerList: List<Logger> = listOf(
-        CommonLogger()
+    private var _loggerList: List<LogWriter> = listOf(
+        CommonWriter()
     )
     private var _defaultTag: String = "Kermit"
     private val lock = Lock()
@@ -29,7 +29,7 @@ internal class LockMutableKermitConfig : MutableKermitConfig {
             }
         }
 
-    override var loggerList: List<Logger>
+    override var loggerList: List<LogWriter>
         get() = _loggerList
         set(value) {
             lock.withLock {
