+++ kermit/src/nativeMain/kotlin/co/touchlab/kermit/KermitConfig.kt
@@ -15,7 +15,7 @@ import kotlin.native.concurrent.freeze
 
 internal class AtomicMutableKermitConfig : MutableKermitConfig {
     private val _minSeverity = AtomicReference(Severity.Debug)
-    private val _loggerList = AtomicReference<List<Logger>>(listOf(CommonLogger()))
+    private val _loggerList = AtomicReference<List<LogWriter>>(listOf(CommonWriter()))
     private val _defaultTag = AtomicReference<String>("Kermit")
 
     override var minSeverity: Severity
@@ -23,7 +23,7 @@ internal class AtomicMutableKermitConfig : MutableKermitConfig {
         set(value) {
             _minSeverity.value = value
         }
-    override var loggerList: List<Logger>
+    override var loggerList: List<LogWriter>
         get() = _loggerList.value
         set(value) {
             _loggerList.value = value.freeze()
